requirements.txt
fastapi
sqlalchemy
uvicorn
python-jose
passlib[bcrypt]